+++ focus-android/app/src/main/java/org/mozilla/focus/session/IntentProcessor.kt
@@ -7,6 +7,7 @@ package org.mozilla.focus.session
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.provider.Browser
 import android.text.TextUtils
 import mozilla.components.browser.state.state.SessionState
 import mozilla.components.feature.customtabs.createCustomTabConfigFromIntent
@@ -61,6 +62,25 @@ class IntentProcessor(
      */
     fun handleNewIntent(context: Context, intent: SafeIntent) {
         createSessionFromIntent(context, intent)
+    }
+
+
+    private fun getAdditionalHeaders(intent: SafeIntent): Map<String, String>? {
+        val pairs = intent.getBundleExtra(Browser.EXTRA_HEADERS)
+        val headers = mutableMapOf<String, String>()
+        pairs?.keySet()?.forEach { key ->
+            val header = pairs.getString(key)
+            if (header != null) {
+                headers[key] = header
+            } else {
+                throw IllegalArgumentException("getAdditionalHeaders() intent bundle contains wrong key value pair")
+            }
+        }
+        return if (headers.isEmpty()) {
+            null
+        } else {
+            headers
+        }
     }
 
     @Suppress("ComplexMethod", "ReturnCount")
@@ -157,6 +177,7 @@ class IntentProcessor(
                     createCustomTabConfigFromIntent(intent.unsafe, context.resources),
                     private = true,
                     source = source,
+                    additionalHeaders = getAdditionalHeaders(intent)
                 ),
             )
         } else {
@@ -166,6 +187,7 @@ class IntentProcessor(
                     source = source,
                     selectTab = true,
                     private = true,
+                    additionalHeaders = getAdditionalHeaders(intent)
                 ),
             )
         }
@@ -183,6 +205,7 @@ class IntentProcessor(
                 createCustomTabConfigFromIntent(intent.unsafe, context.resources),
                 private = true,
                 source = source,
+                additionalHeaders = getAdditionalHeaders(intent)
             )
             Pair(Result.CustomTab(tabId), tabId)
         } else {
@@ -190,6 +213,7 @@ class IntentProcessor(
                 url,
                 source = source,
                 private = true,
+                additionalHeaders = getAdditionalHeaders(intent)
             )
             Pair(Result.Tab(tabId), tabId)
         }